+++ BV3N92_HFT_2021221.Logic/PartyLogic.cs
@@ -19,22 +19,87 @@ namespace BV3N92_HFT_2021221.Logic
 
         public void ChangeIdeology(int partyId, Ideologies newIdeology)
         {
-            partyRepo.ChangeIdeology(partyId, newIdeology);
+            if (partyId.Equals(null))
+            {
+                throw new Exception("ID has to be given!");
+            }
+            else if (GetPartyByID(partyId).Ideology.Equals(newIdeology))
+            {
+                throw new Exception("New ideology cannot match old one!");
+            }
+            else
+                partyRepo.ChangeIdeology(partyId, newIdeology);
         }
 
         public void ChangePartyName(int partyId, string newName)
         {
-            partyRepo.ChangePartyName(partyId, newName);
+            if (partyId.Equals(null))
+            {
+                throw new Exception("ID has to be given!");
+            }
+            else if (newName.Equals(string.Empty))
+            {
+                throw new Exception("Name has to be given!");
+            }
+            else if (GetPartyByID(partyId).PartyName.Equals(newName))
+            {
+                throw new Exception("New name cannot match old one!");
+            }
+            else
+                partyRepo.ChangePartyName(partyId, newName);
         }
 
         public void CreateParty(int partyId, int parliamentId, string partyName, Ideologies ideology)
         {
-            partyRepo.CreateParty(partyId, parliamentId, partyName, ideology);
+            foreach (var item in GetAllParties())
+            {
+                if (item.PartyName.Equals(partyName))
+                {
+                    throw new Exception($"A party with the name '{partyName}' already exists!");
+                }
+                else if (item.PartyID.Equals(partyId))
+                {
+                    throw new Exception($"A party with ID '{partyId}' already exists!");
+                }
+            }
+
+            if (partyId.Equals(null))
+            {
+                throw new Exception("Party ID has to be given!");
+            }
+            else if (parliamentId.Equals(null))
+            {
+                throw new Exception("Parliament ID has to be given!");
+            }
+            else if (partyName.Equals(string.Empty))
+            {
+                throw new Exception("Name has to be given!");
+            }
+            else
+                partyRepo.CreateParty(partyId, parliamentId, partyName, ideology);
         }
 
         public void DeleteParty(int partyId)
         {
-            partyRepo.DeleteParty(partyId);
+            int i = 0;
+            foreach (var item in GetAllParties())
+            {
+                if (item.PartyID.Equals(partyId))
+                {
+                    i++;
+                }
+            }
+
+            if (partyId.Equals(null))
+            {
+                throw new Exception("ID has to be given!");
+            }
+            else if (i == 0)
+            {
+                throw new Exception($"No such party with ID '{partyId}'!");
+            }
+            else
+                partyRepo.DeleteParty(partyId);
         }
 
         public IList<Party> GetAllParties()
@@ -44,7 +109,24 @@ namespace BV3N92_HFT_2021221.Logic
 
         public Party GetPartyByID(int id)
         {
-            return partyRepo.GetOne(id);
+            int i = 0;
+            foreach (var item in GetAllParties())
+            {
+                if (item.PartyID.Equals(id))
+                {
+                    i++;
+                }
+            }
+            if (id.Equals(null))
+            {
+                throw new Exception("ID has to be given!");
+            }
+            else if (i == 0)
+            {
+                throw new Exception($"No such party with ID '{id}'!");
+            }
+            else
+                return partyRepo.GetOne(id);
         }
     }
 }
